Move db aws stat reporter to external file.
